+++ entity-view/processor/src/test/java/com/blazebit/persistence/view/processor/model/BView.java
@@ -1,6 +1,7 @@
 package com.blazebit.persistence.view.processor.model;
 
 import com.blazebit.persistence.view.EntityView;
+import com.blazebit.persistence.view.PostLoad;
 import com.blazebit.persistence.view.Self;
 import com.blazebit.persistence.view.ViewConstructor;
 import com.blazebit.persistence.view.processor.model.sub.BaseView;
@@ -11,6 +12,7 @@ import java.io.Serializable;
 public abstract class BView<X extends Serializable> extends BaseView<Integer> {
 
     private final String capturedName;
+    private String postLoadName;
 
     @ViewConstructor("create")
     public BView() {
@@ -21,11 +23,20 @@ public abstract class BView<X extends Serializable> extends BaseView<Integer> {
         this.capturedName = self == null ? null : self.getName();
     }
 
+    @PostLoad
+    void postLoad() {
+        this.postLoadName = getName();
+    }
+
     public abstract String getName();
 
     public abstract void setName(String name);
 
     public String getCapturedName() {
         return capturedName;
+    }
+
+    public String getPostLoadName() {
+        return postLoadName;
     }
 }